move ModelBundle out into its own file
