+++ nucleus/experimental/model_bundle.py
@@ -1,0 +1,8 @@
+class ModelBundle:
+    """
+    Represents a ModelBundle.
+    TODO fill this out with more than just a name potentially.
+    """
+
+    def __init__(self, name):
+        self.name = name
